add LinearAlgebraExt
