+++ Project.toml
@@ -9,20 +9,24 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 [weakdeps]
 IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 
 [extensions]
 ConstructionBaseIntervalSetsExt = "IntervalSets"
 ConstructionBaseStaticArraysExt = "StaticArrays"
+ConstructionBaseLinearAlgebraExt = "LinearAlgebra"
 
 [compat]
 IntervalSets = "0.5, 0.6, 0.7"
 StaticArrays = "1"
 julia = "1"
+LinearAlgebra = "1"
 
 [extras]
 IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 
 [targets]
-test = ["IntervalSets","StaticArrays","Test"]
+test = ["IntervalSets","LinearAlgebra","StaticArrays","Test"]
