+++ src/nonstandard.jl
@@ -28,24 +28,7 @@ end
 constructorof(::Type{<:PermutedDimsArray{<:Any,N,perm,iperm,<:Any}}) where {N,perm,iperm} =
     PermutedDimsArrayConstructor{N,perm,iperm}()
 
-### Tridiagonal
-function tridiagonal_constructor(dl::V, d::V, du::V) where {V<:AbstractVector{T}} where T
-    Tridiagonal{T,V}(dl, d, du)
-end
-function tridiagonal_constructor(dl::V, d::V, du::V, du2::V) where {V<:AbstractVector{T}} where T
-    Tridiagonal{T,V}(dl, d, du, du2)
-end
 
-# `du2` may be undefined, so we need a custom `getfields` that checks `isdefined`
-function getfields(o::Tridiagonal)
-    if isdefined(o, :du2)
-        (dl=o.dl, d=o.d, du=o.du, du2=o.du2) 
-    else
-        (dl=o.dl, d=o.d, du=o.du)
-    end
-end
-
-constructorof(::Type{<:LinearAlgebra.Tridiagonal}) = tridiagonal_constructor
 
 ### LinRange
 # `lendiv` is a calculated field
@@ -57,20 +40,6 @@ constructorof(::Type{<:LinRange}) = linrange_constructor
 # ::Expr annotation is to make it type-stable on Julia 1.3-
 constructorof(::Type{<:Expr}) = (head, args) -> Expr(head, args...)::Expr
 
-### Cholesky
-setproperties(C::LinearAlgebra.Cholesky, patch::NamedTuple{()}) = C
-function setproperties(C::LinearAlgebra.Cholesky, patch::NamedTuple{(:L,),<:Tuple{<:LinearAlgebra.LowerTriangular}})
-    return LinearAlgebra.Cholesky(C.uplo === 'U' ? copy(patch.L.data') : patch.L.data, C.uplo, C.info)
-end
-function setproperties(C::LinearAlgebra.Cholesky, patch::NamedTuple{(:U,),<:Tuple{<:LinearAlgebra.UpperTriangular}})
-    return LinearAlgebra.Cholesky(C.uplo === 'L' ? copy(patch.U.data') : patch.U.data, C.uplo, C.info)
-end
-function setproperties(
-    C::LinearAlgebra.Cholesky,
-    patch::NamedTuple{(:UL,),<:Tuple{<:Union{LinearAlgebra.LowerTriangular,LinearAlgebra.UpperTriangular}}}
-)
-    return LinearAlgebra.Cholesky(patch.UL.data, C.uplo, C.info)
-end
-function setproperties(C::LinearAlgebra.Cholesky, patch::NamedTuple)
-    throw(ArgumentError("Invalid patch for `Cholesky`: $(patch)"))
+if !isdefined(Base,:get_extension)
+    include("ext/ConstructionBaseLinearAlgebraExt.jl")
 end
